refactor: 팀 규칙으로 인해 hooks 디렉토리로 훅 위치 수정 #171
